anti oishing/requirements.txt
streamlit
pandas
scikit-learn
python-whois
requests
